requirements.txt
fastapi
uvicorn
google-adk
google-generativeai
dotenv
